refactor: update unit tests
