+++ src/table/hooks/__tests__/use-snapshot.spec.ts
@@ -1,77 +1,20 @@
-import { generateLayout } from '../../../__test__/generate-test-data';
+import * as stardust from '@nebula.js/stardust';
 import { TableLayout, ViewService } from '../../../types';
-import { getVisibleHeight, getViewState } from '../use-snapshot';
-import * as visibleRowsUtils from '../../utils/find-visible-rows';
-import * as handleData from '../../../handle-data';
+import * as snapshotWrapper from '../use-snapshot';
+import * as getViewStateWrapper from '../../utils/get-view-state';
+
+jest.mock('@nebula.js/stardust', () => ({
+  onTakeSnapshot: jest.fn(),
+  useImperativeHandle: jest.fn(),
+}));
 
 describe('use-snapshot', () => {
-  describe('getVisibleHeight', () => {
-    let visibleRowEndIndex: number;
-    let visibleRowStartIndex: number;
-    let layout: TableLayout;
-    let viewService: ViewService;
-    let rowCount: number;
-    const dimensionCount = 2;
-    const measureCount = 1;
-
-    beforeEach(() => {
-      visibleRowStartIndex = 0;
-      visibleRowEndIndex = 79;
-      viewService = {
-        scrollTopRatio: 1,
-        visibleTop: 0,
-        visibleHeight: 80,
-        page: 0,
-        rowsPerPage: 100,
-        qTop: 0,
-        qHeight: 100,
-        estimatedRowHeight: 25
-      };
-      rowCount = 50;
-    });
-
-    it('should check if qcy is used', async () => {
-      layout = generateLayout(dimensionCount, measureCount, rowCount);
-
-      const result = getVisibleHeight(visibleRowEndIndex, visibleRowStartIndex, layout, viewService);
-      expect(result).toBe(rowCount);
-    });
-
-    it('should check if rowsPerPage is used', async () => {
-      visibleRowEndIndex = 13;
-      layout = generateLayout(dimensionCount, measureCount, rowCount);
-      viewService = {
-        ...viewService,
-        visibleHeight: 14,
-        rowsPerPage: 10,
-      };
-
-      const result = getVisibleHeight(visibleRowEndIndex, visibleRowStartIndex, layout, viewService);
-      expect(result).toBe(viewService.rowsPerPage);
-    });
-
-    it('should add extra rows when row count is larger than rowsPerPage', async () => {
-      visibleRowEndIndex = 5;
-      layout = generateLayout(dimensionCount, measureCount, rowCount);
-      viewService = { ...viewService, visibleHeight: 6, rowsPerPage: 10 };
-
-      const result = getVisibleHeight(visibleRowEndIndex, visibleRowStartIndex, layout, viewService);
-      expect(result).toBe(9);
-    });
-
-    it('should return 0 when visibleRowEndIndex < 0', async () => {
-      visibleRowEndIndex = -1;
-      layout = generateLayout(dimensionCount, measureCount, rowCount);
-
-      const result = getVisibleHeight(visibleRowEndIndex, visibleRowStartIndex, layout, viewService);
-      expect(result).toBe(0);
-    });
-  });
-
-  describe('getViewState', () => {
+  describe('useSnapshot', () => {
     let layout: TableLayout;
     let viewService: ViewService;
     let rootElement: HTMLElement;
+    let model: undefined;
+    let contentRect: stardust.stardust.Rect;
 
     beforeEach(() => {
       layout = {} as TableLayout;
@@ -87,53 +30,38 @@ describe('use-snapshot', () => {
         estimatedRowHeight: 25,
       };
       rootElement = document.createElement('p');
-      jest.spyOn(visibleRowsUtils, 'findVirtualizedVisibleRows').mockReturnValue({});
-      jest.spyOn(visibleRowsUtils, 'findPaginationVisibleRows').mockReturnValue({});
-      jest.spyOn(handleData, 'getTotalPosition').mockReturnValue({ atTop: false, atBottom: false });
+      model = undefined;
+      contentRect = { top: 0, left: 0, width: 100, height: 100 };
+      jest.spyOn(getViewStateWrapper, 'default').mockReturnValue({
+        scrollLeft: 1,
+        visibleLeft: 2,
+        visibleWidth: 3,
+        visibleTop: 4,
+        visibleHeight: 5,
+        scrollTopRatio: 6,
+        rowsPerPage: 7,
+        page: 0,
+      });
     });
 
     afterEach(() => {
       jest.resetAllMocks();
     });
 
-    it('should run correct functions in virtualized table mode', async () => {
-      getViewState(layout, viewService, rootElement);
-      expect(handleData.getTotalPosition).toHaveBeenCalledTimes(0);
-      expect(visibleRowsUtils.findPaginationVisibleRows).toHaveBeenCalledTimes(0);
-      expect(visibleRowsUtils.findVirtualizedVisibleRows).toHaveBeenCalledTimes(1);
-    });
-
-    it('should return correct result in virtualized table mode', async () => {
-      const result = getViewState(layout, viewService, rootElement);
+    it('useImperativeHandle should run correct getViewState', async () => {
+      snapshotWrapper.default({ layout, viewService, model, rootElement, contentRect });
+      const { getViewState } = (stardust.useImperativeHandle as jest.Mock).mock.calls[0][0]();
+      const result = getViewState();
+      expect(getViewStateWrapper.default).toHaveBeenCalledTimes(1);
+      expect(getViewStateWrapper.default).toHaveBeenCalledWith(layout, viewService, rootElement);
       expect(result).toEqual({
-        scrollLeft: 100,
-        visibleLeft: 1,
-        visibleWidth: 10,
-        visibleTop: -1,
-        visibleHeight: 0,
-        scrollTopRatio: 1,
-        rowsPerPage: 100,
-        page: 0,
-      });
-    });
-
-    it('should run correct functions in pagination table mode', async () => {
-      layout.usePagination = true;
-      getViewState(layout, viewService, rootElement);
-      expect(handleData.getTotalPosition).toHaveBeenCalledTimes(1);
-      expect(visibleRowsUtils.findPaginationVisibleRows).toHaveBeenCalledTimes(1);
-      expect(visibleRowsUtils.findVirtualizedVisibleRows).toHaveBeenCalledTimes(0);
-    });
-
-    it('should return correct result in pagination table mode', async () => {
-      layout.usePagination = true;
-      const result = getViewState(layout, viewService, rootElement);
-      expect(result).toEqual({
-        rowPartialHeight: undefined,
-        scrollLeft: 100,
-        visibleTop: -1,
-        visibleHeight: 0,
-        rowsPerPage: 100,
+        scrollLeft: 1,
+        visibleLeft: 2,
+        visibleWidth: 3,
+        visibleTop: 4,
+        visibleHeight: 5,
+        scrollTopRatio: 6,
+        rowsPerPage: 7,
         page: 0,
       });
     });
